refactor: split pattern helpers and cleanup
