+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/constants.ts
@@ -8,7 +8,7 @@
 import { EcsFlat } from '@elastic/ecs';
 import { EuiComboBoxOptionOption } from '@elastic/eui';
 
-import { EcsFieldMetadata } from './types';
+import { EcsFieldMetadata, SortConfig } from './types';
 import * as i18n from './translations';
 
 export const EcsFlatTyped = EcsFlat as unknown as Record<string, EcsFieldMetadata>;
@@ -42,3 +42,10 @@ export const ilmPhaseOptionsStatic: EuiComboBoxOptionOption[] = [
 export const EMPTY_STAT = '--';
 
 export const INTERNAL_API_VERSION = '1';
+
+export const defaultSort: SortConfig = {
+  sort: {
+    direction: 'desc',
+    field: 'docsCount',
+  },
+};
