+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/index.tsx
@@ -9,15 +9,6 @@ import { EuiSpacer, useGeneratedHtmlId } from '@elastic/eui';
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
 import { ErrorEmptyPrompt } from './error_empty_prompt';
-import {
-  defaultSort,
-  getIlmExplainPhaseCounts,
-  getPageIndex,
-  getSummaryTableItems,
-  MIN_PAGE_SIZE,
-  shouldCreateIndexNames,
-  shouldCreatePatternRollup,
-} from './helpers';
 import { getTotalPatternIncompatible, getTotalPatternIndicesChecked } from '../../../utils/stats';
 import { getIndexNames, getPatternDocsCount, getPatternSizeInBytes } from './utils/stats';
 import { LoadingEmptyPrompt } from './loading_empty_prompt';
@@ -34,6 +25,13 @@ import { PatternAccordion, PatternAccordionChildren } from './styles';
 import { IndexCheckFlyout } from './index_check_flyout';
 import { useResultsRollupContext } from '../../../contexts/results_rollup_context';
 import { useIndicesCheckContext } from '../../../contexts/indices_check_context';
+import { getSummaryTableItems } from '../../../utils/get_summary_table_items';
+import { defaultSort } from '../../../constants';
+import { MIN_PAGE_SIZE } from './constants';
+import { getIlmExplainPhaseCounts } from './utils/ilm_explain';
+import { shouldCreateIndexNames } from './utils/should_create_index_names';
+import { shouldCreatePatternRollup } from './utils/should_create_pattern_rollup';
+import { getPageIndex } from './utils/get_page_index';
 
 const EMPTY_INDEX_NAMES: string[] = [];
