+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/summary_table/helpers.tsx
@@ -18,47 +18,21 @@ import moment from 'moment';
 import styled from 'styled-components';
 
 import { euiThemeVars } from '@kbn/ui-theme';
+import { IndexSummaryTableItem } from '../../../../types';
 import { EMPTY_STAT } from '../../../../constants';
 import { getIlmPhaseDescription } from '../../../../utils/get_ilm_phase_description';
 import { INCOMPATIBLE_INDEX_TOOL_TIP } from '../../../../stat_label/translations';
 import { INDEX_SIZE_TOOLTIP } from '../../../../translations';
 import * as i18n from './translations';
-import { IndexSummaryTableItem } from '../types';
 import { UseIndicesCheckCheckState } from '../../../../hooks/use_indices_check/types';
 import { IndexResultBadge } from '../index_result_badge';
 import { getIndexResultToolTip } from '../index_result_badge/helpers';
 import { Stat } from '../../../../stat';
+import { getDocsCountPercent } from '../utils/stats';
 
 const ProgressContainer = styled.div`
   width: 150px;
 `;
-
-export const getDocsCountPercent = ({
-  docsCount,
-  locales,
-  patternDocsCount,
-}: {
-  docsCount: number;
-  locales?: string | string[];
-  patternDocsCount: number;
-}): string =>
-  patternDocsCount !== 0
-    ? Number(docsCount / patternDocsCount).toLocaleString(locales, {
-        style: 'percent',
-        maximumFractionDigits: 1,
-        minimumFractionDigits: 1,
-      })
-    : '';
-
-export const getToggleButtonId = ({
-  indexName,
-  isExpanded,
-  pattern,
-}: {
-  indexName: string;
-  isExpanded: boolean;
-  pattern: string;
-}): string => (isExpanded ? `collapse${indexName}${pattern}` : `expand${indexName}${pattern}`);
 
 export const getSummaryTableILMPhaseColumn = (
   isILMAvailable: boolean
