+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/storage_details/utils/get_flattened_buckets.ts
@@ -6,8 +6,8 @@
  */
 
 import { PatternRollup } from '../../../types';
+import { getIlmPhase } from '../../../utils/get_ilm_phase';
 import { getDocsCount, getSizeInBytes } from '../../../utils/stats';
-import { getIlmPhase } from '../../indices_details/pattern/helpers';
 import { FlattenedBucket } from '../types';
 
 export const getFlattenedBuckets = ({
